+++ src/ga/FitnessEvaluator.java
@@ -3,7 +3,9 @@ package ga;
 import ga.constraints.ConsecutiveLectureConstraint;
 import ga.constraints.GapConstraint;
 import ga.constraints.MaxDailyLectureConstraint;
+import ga.constraints.OverlapConstraint;
 import ga.constraints.SpreadConstraint;
+import ga.constraints.TutorialBeforeLabConstraint;
 import ga.constraints.WeightedConstraint;
 
 import java.util.ArrayList;
@@ -15,10 +17,12 @@ public class FitnessEvaluator {
 
     public FitnessEvaluator() {
 
+        constraints.add(new WeightedConstraint(new OverlapConstraint(), 100));
         constraints.add(new WeightedConstraint(new ConsecutiveLectureConstraint(), 10));
         constraints.add(new WeightedConstraint(new SpreadConstraint(), 10));
         constraints.add(new WeightedConstraint(new MaxDailyLectureConstraint(), 10));
         constraints.add(new WeightedConstraint(new GapConstraint(), 5));
+        constraints.add(new WeightedConstraint(new TutorialBeforeLabConstraint(), 10));
     }
 
     public int evaluate(Chromosome chromosome) {